// face-check-ui/src/app/services/models/work-site-all-information-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LocalTime } from '../models/local-time';
export interface WorkSiteAllInformationResponse {
  address?: string;
  allowedRadius?: number;
  isActive?: boolean;
  latitude?: number;
  longitude?: number;
  siteName?: string;
  workDayEnd?: LocalTime;
  workDayStart?: LocalTime;
  workSiteId?: number;
}